system/requirements.txt
# SAEC - Dependências
# Instalar com: pip install -r requirements.txt

# Core
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.0

# PDF e Imagens
pymupdf>=1.24.0
pillow>=10.0

# LLM APIs
anthropic>=0.30.0
openai>=1.30.0
httpx>=0.27.0
ollama>=0.4.0

# Notebooks
ipywidgets>=8.0
ipython>=8.0
jupyter>=1.0.0

# Consolidação
openpyxl>=3.1.0
pandas>=2.0
rapidfuzz>=3.9.0

# Utilitários
rich>=13.0
tqdm>=4.66.0
chromadb>=0.5.5

# Linting e Type Checking
pylint>=4.0.0
mypy>=1.10.0
pyright>=1.1.400
types-PyYAML>=6.0.12

# Tests
pytest>=8.0.0
pytest-mock>=3.12.0

